Add automatic version detection
This is more of a base to work off of. Unfortunately adds dependency on `jq` and `curl`. I don't know FreeBSD or pfSense, but I noticed `fetch` was used to download the file. Maybe that could replace `curl`? The version fetching works great on Linux. Some other changes will need to be made, because you can no longer assume the version is valid.